+++ ingress-controller/internal/dataplane/kong_client.go
@@ -198,6 +198,14 @@ type KongClient struct {
 	// konnectKongStateUpdater is used to update the current state seen by Konnect that will be picked asynchronously
 	// by the Konnect config synchronization loop.
 	konnectKongStateUpdater KonnectKongStateUpdater
+
+	// hasSuccessfullyPushedConfig indicates whether this client instance has already
+	// successfully pushed a configuration to a gateway during its lifetime.
+	// It is used to prevent an initial empty config push from replacing a known
+	// last valid configuration during startup race between last valid config fetching
+	// and syncing informer cache.
+	hasSuccessfullyPushedConfig     bool
+	hasSuccessfullyPushedConfigOnce sync.Once
 }
 
 // KongClientOption is a functional option for configuring a KongClient.
@@ -515,19 +523,27 @@ func (c *KongClient) Update(ctx context.Context) error {
 	parsingResult := c.kongConfigBuilder.BuildKongConfig()
 	translationDuration := time.Since(translationStart)
 
-	if failuresCount := len(parsingResult.TranslationFailures); failuresCount > 0 {
-		c.metricsRecorder.RecordTranslationFailure(translationDuration)
-		c.metricsRecorder.RecordTranslationBrokenResources(failuresCount)
-		c.recordResourceFailureEvents(parsingResult.TranslationFailures, KongConfigurationTranslationFailedEventReason)
-		c.logger.V(logging.DebugLevel).Info("Translation failures occurred when building data-plane configuration", "count", failuresCount)
+	kongState := parsingResult.KongState
+	if newKongState, skip := c.shouldSkipInitialEmptyConfigPush(parsingResult.KongState); skip {
+		// Use last valid state if we should skip the initial empty config.
+		kongState = newKongState
 	} else {
-		c.metricsRecorder.RecordTranslationSuccess(translationDuration)
-		c.metricsRecorder.RecordTranslationBrokenResources(0)
-		c.logger.V(logging.DebugLevel).Info("Successfully built data-plane configuration", "duration", translationDuration.String())
+		// Otherwise, the translated Kong state is used so we should record
+		// the statistics of the translation.
+		if failuresCount := len(parsingResult.TranslationFailures); failuresCount > 0 {
+			c.metricsRecorder.RecordTranslationFailure(translationDuration)
+			c.metricsRecorder.RecordTranslationBrokenResources(failuresCount)
+			c.recordResourceFailureEvents(parsingResult.TranslationFailures, KongConfigurationTranslationFailedEventReason)
+			c.logger.V(logging.DebugLevel).Info("Translation failures occurred when building data-plane configuration", "count", failuresCount)
+		} else {
+			c.metricsRecorder.RecordTranslationSuccess(translationDuration)
+			c.metricsRecorder.RecordTranslationBrokenResources(0)
+			c.logger.V(logging.DebugLevel).Info("Successfully built data-plane configuration", "duration", translationDuration.String())
+		}
 	}
 
 	const isFallback = false
-	shas, gatewaysSyncErr := c.sendOutToGatewayClients(ctx, parsingResult.KongState, c.kongConfig, isFallback)
+	shas, gatewaysSyncErr := c.sendOutToGatewayClients(ctx, kongState, c.kongConfig, isFallback)
 
 	// Taking into account the results of syncing configuration with Gateways and potential translation
 	// failures, calculate the config status and update it.
@@ -551,7 +567,7 @@ func (c *KongClient) Update(ctx context.Context) error {
 	}
 
 	// Send configuration to Konnect only when successfully applied configuration to Kong Gateways run in cluster.
-	c.maybeUpdateKonnectKongState(parsingResult.KongState, isFallback)
+	c.maybeUpdateKonnectKongState(kongState, isFallback)
 	// Gateways were successfully synced with the current configuration, so we can update the last valid cache snapshot.
 	c.maybePreserveTheLastValidConfigCache(cacheSnapshot)
 
@@ -577,6 +593,35 @@ func (c *KongClient) maybePreserveTheLastValidConfigCache(lastValidCache store.C
 		c.logger.V(logging.DebugLevel).Info("Preserving the last valid configuration cache")
 		c.lastValidCacheSnapshot = &lastValidCache
 	}
+}
+
+// shouldSkipInitialEmptyConfigPush returns true if the client should skip pushing
+// an initial empty configuration to the gateways, which can happen during startup
+// when the informer cache has not yet been fully populated and the client has
+// not yet successfully pushed any configuration.
+// In this case, if there is a last valid configuration available,
+// we can skip pushing an empty configuration and instead use the last valid
+// configuration as the one to be pushed, which can help avoid unnecessary downtime
+// due to pushing an empty configuration.
+// Last valid configuration is returned when function returns true.
+func (c *KongClient) shouldSkipInitialEmptyConfigPush(s *kongstate.KongState) (*kongstate.KongState, bool) {
+	if c.hasSuccessfullyPushedConfig {
+		return nil, false
+	}
+
+	if !s.IsEmpty() {
+		return nil, false
+	}
+
+	lastValidConfig, ok := c.kongConfigFetcher.LastValidConfig()
+	if !ok || lastValidConfig == nil || lastValidConfig.IsEmpty() {
+		return nil, false
+	}
+
+	c.logger.V(logging.DebugLevel).Info(
+		"Skipping empty config push before first successful update because a last valid config is already available",
+	)
+	return lastValidConfig, true
 }
 
 // maybeTryRecoveringFromGatewaysSyncError tries to recover from a configuration rejection if the error is of the expected
@@ -771,6 +816,9 @@ func (c *KongClient) sendOutToGatewayClients(
 	previousSHAs := c.SHAs
 	sort.Strings(shas)
 	c.SHAs = shas
+	c.hasSuccessfullyPushedConfigOnce.Do(func() {
+		c.hasSuccessfullyPushedConfig = true
+	})
 
 	c.kongConfigFetcher.StoreLastValidConfig(s)