+++ ingress-controller/internal/konnect/node_agent_test.go
@@ -21,7 +21,8 @@ import (
 )
 
 const (
-	testHostname = "ingress-0"
+	testHostname             = "ingress-0"
+	defaultRefreshNodePeriod = 60 * time.Second
 )
 
 var (
@@ -274,7 +275,7 @@ func TestNodeAgentUpdateNodes(t *testing.T) {
 			nodeAgent := konnect.NewNodeAgent(
 				testHostname,
 				testKOUserAgent,
-				konnect.DefaultRefreshNodePeriod,
+				defaultRefreshNodePeriod,
 				logr.Discard(),
 				nodeClient,
 				configStatusQueue,
@@ -342,7 +343,7 @@ func TestNodeAgent_StartDoesntReturnUntilContextGetsCancelled(t *testing.T) {
 	nodeAgent := konnect.NewNodeAgent(
 		testHostname,
 		testKOUserAgent,
-		konnect.DefaultRefreshNodePeriod,
+		defaultRefreshNodePeriod,
 		logr.Discard(),
 		nodeClient,
 		newMockConfigStatusNotifier(),
@@ -387,7 +388,7 @@ func TestNodeAgent_ControllerNodeStatusGetsUpdatedOnStatusNotification(t *testin
 	nodeAgent := konnect.NewNodeAgent(
 		testHostname,
 		testKOUserAgent,
-		konnect.DefaultRefreshNodePeriod,
+		defaultRefreshNodePeriod,
 		logr.Discard(),
 		nodeClient,
 		configStatusQueue,
@@ -489,7 +490,7 @@ func TestNodeAgent_ControllerNodeStatusGetsUpdatedOnlyWhenItChanges(t *testing.T
 	nodeAgent := konnect.NewNodeAgent(
 		testHostname,
 		testKOUserAgent,
-		konnect.DefaultRefreshNodePeriod,
+		defaultRefreshNodePeriod,
 		logr.Discard(),
 		nodeClient,
 		configStatusQueue,
@@ -528,7 +529,7 @@ func TestNodeAgent_ControllerNodeStatusGetsUpdatedOnlyWhenItChanges(t *testing.T
 }
 
 func TestNodeAgent_TickerResetsOnEveryNodesUpdate(t *testing.T) {
-	const halfOfRefreshPeriod = konnect.DefaultRefreshNodePeriod / 2
+	const halfOfRefreshPeriod = defaultRefreshNodePeriod / 2
 
 	t.Run("config status notification", func(t *testing.T) {
 		nodeClient := newMockNodeClient(nil)
@@ -539,7 +540,7 @@ func TestNodeAgent_TickerResetsOnEveryNodesUpdate(t *testing.T) {
 		nodeAgent := konnect.NewNodeAgent(
 			testHostname,
 			testKOUserAgent,
-			konnect.DefaultRefreshNodePeriod,
+			defaultRefreshNodePeriod,
 			logr.Discard(),
 			nodeClient,
 			configStatusQueue,
@@ -563,7 +564,7 @@ func TestNodeAgent_TickerResetsOnEveryNodesUpdate(t *testing.T) {
 		require.Eventually(t, func() bool { return nodeClient.NodesUpdatesCount() == 2 }, time.Second, time.Microsecond)
 
 		t.Log("trigger update with ticker")
-		ticker.Add(konnect.DefaultRefreshNodePeriod)
+		ticker.Add(defaultRefreshNodePeriod)
 		require.Eventually(t, func() bool { return nodeClient.NodesUpdatesCount() > 2 }, time.Second, time.Microsecond)
 	})
 
@@ -576,7 +577,7 @@ func TestNodeAgent_TickerResetsOnEveryNodesUpdate(t *testing.T) {
 		nodeAgent := konnect.NewNodeAgent(
 			testHostname,
 			testKOUserAgent,
-			konnect.DefaultRefreshNodePeriod,
+			defaultRefreshNodePeriod,
 			logr.Discard(),
 			nodeClient,
 			configStatusQueue,
@@ -599,7 +600,7 @@ func TestNodeAgent_TickerResetsOnEveryNodesUpdate(t *testing.T) {
 		require.Eventually(t, func() bool { return nodeClient.NodesUpdatesCount() == 2 }, time.Second, time.Microsecond)
 
 		t.Log("trigger update with ticker")
-		ticker.Add(konnect.DefaultRefreshNodePeriod)
+		ticker.Add(defaultRefreshNodePeriod)
 		require.Eventually(t, func() bool { return nodeClient.NodesUpdatesCount() == 3 }, time.Second, time.Microsecond)
 	})
 }